Create types files locally in each panel
