+++ src/panels/MissionsPanel/MissionContent.tsx
@@ -3,11 +3,10 @@ import { Button, Group, ScrollArea, Switch, Text, Title } from '@mantine/core';
 
 import { useSubscribeToTime } from '@/api/hooks';
 import { InfoBox } from '@/components/InfoBox/InfoBox';
-import { DisplayType } from '@/types/enums';
-import { DisplayedPhase, Phase } from '@/types/mission-types';
 
 import { TimeLine } from './TimeLine/TimeLine';
 import { MissionPhase } from './MissionPhase';
+import { DisplayedPhase, DisplayType, Phase } from './types';
 
 interface Props {
   missionOverview: Phase;
