+++ src/panels/TimePanel/SimulationIncrement.tsx
@@ -4,11 +4,10 @@ import { useThrottledCallback } from '@mantine/hooks';
 
 import { useOpenSpaceApi } from '@/api/hooks';
 import { useAppSelector } from '@/redux/hooks';
-import { TimePart } from '@/types/enums';
 
 import { DeltaTimeStepsControl } from './DeltaTimeStepControl';
 import { TimeIncrementInput } from './TimeIncrementInput';
-import { Decimals, StepSizes } from './util';
+import { Decimals, StepSizes, TimePart } from './types';
 
 export function SimulationIncrement() {
   const updateDeltaTime = useThrottledCallback(updateDeltaTimeNow, 50);
