+++ src/redux/documentation/documentationMiddleware.ts
@@ -3,9 +3,9 @@ import { createAsyncThunk } from '@reduxjs/toolkit';
 import { api } from '@/api/api';
 import { onOpenConnection } from '@/redux/connection/connectionSlice';
 import { AppStartListening } from '@/redux/listenerMiddleware';
-import { AssetMetaData } from '@/types/types';
 
 import { initializeDocumentation } from './documentationSlice';
+import { AssetMetaData } from './types';
 
 export const fetchDocumentation = createAsyncThunk(
   'documentation/fetchDocumentation',
