+++ src/redux/documentation/types.ts
@@ -1,0 +1,10 @@
+export interface AssetMetaData {
+  author: string;
+  description: string;
+  identifiers?: string[];
+  license: string;
+  name: string;
+  path: string;
+  url: string;
+  version: string;
+}
