+++ src/redux/time/types.ts
@@ -1,0 +1,12 @@
+export interface OpenSpaceTimeState {
+  time?: string;
+  timeCapped?: string;
+  targetDeltaTime?: number;
+  deltaTime?: number;
+  isPaused?: boolean;
+  hasNextStep?: boolean;
+  hasPrevStep?: boolean;
+  nextStep?: number;
+  prevStep?: number;
+  deltaTimeSteps?: number[];
+}
