+++ src/types/types.ts
@@ -1,9 +1,12 @@
 import { Property } from '@/components/Property/Property';
+
 import { PropertyVisibilityNumber } from './enums';
 
+/*--------------------------------------------------------------------------------------*/
 export type Uri = string;
 export type Identifier = string;
 
+/*--------------------------------------------------------------------------------------*/
 export interface Action {
   identifier: string;
   guiPath: string;
@@ -20,45 +23,14 @@ export interface Keybind {
 
 export type ActionOrKeybind = Action | Keybind;
 
+/*--------------------------------------------------------------------------------------*/
 export interface SemanticVersion {
   major: number;
   minor: number;
   patch: number;
 }
 
-// Incomplete type objects for the ArcGIS object we request for GeoLocationPanel
-// This was to avoid having to import the entire ArcGIS core since the @types file is
-// deprecated on npm.
-export type ArcGISJSON = {
-  candidates: Candidate[];
-  spatialReference: object;
-};
-
-export type Candidate = {
-  address: string;
-  attributes: {
-    LongLabel: string;
-  };
-  extent: Extent;
-  location: Location;
-};
-
-export type Location = {
-  x: number;
-  y: number;
-};
-
-export type Extent = {
-  xmax: number;
-  xmin: number;
-  ymax: number;
-  ymin: number;
-};
-
-export interface ExoplanetData {
-  name: string;
-  identifier: Identifier;
-}
+/*--------------------------------------------------------------------------------------*/
 
 export type PropertyVisibility = keyof typeof PropertyVisibilityNumber;
 
@@ -110,6 +82,7 @@ export interface PropertyOwners {
   [key: Uri]: PropertyOwner | undefined;
 }
 
+/*--------------------------------------------------------------------------------------*/
 export type Group = {
   subgroups: string[]; // group paths
   propertyOwners: Uri[];
@@ -119,31 +92,8 @@ export type Groups = {
   [key: string]: Group;
 };
 
+/*--------------------------------------------------------------------------------------*/
 export type CustomGroupOrdering = {
   // The value is a list of node names in the order they should be displayed
   [key: string]: string[] | undefined;
-};
-
-export interface AssetMetaData {
-  author: string;
-  description: string;
-  identifiers?: string[];
-  license: string;
-  name: string;
-  path: string;
-  url: string;
-  version: string;
-}
-
-export type OpenSpaceTimeState = {
-  time?: string;
-  timeCapped?: string;
-  targetDeltaTime?: number;
-  deltaTime?: number;
-  isPaused?: boolean;
-  hasNextStep?: boolean;
-  hasPrevStep?: boolean;
-  nextStep?: number;
-  prevStep?: number;
-  deltaTimeSteps?: number[];
 };
